snowflake/mcp/fraud-analysis-mcp-implementation/requirements.txt
aiohttp
httpx
mcp[cli]
boto3
dotenv
pyyaml
strands-agents
strands-agents-tools